♻️ Extraced JahrgangInput.vue from NwkCreateDialog.vue and moved <template> in NameInput to the top
